more file and class renaming -- seems better now
